requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
Brotli==1.0.9
celery==5.3.1
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==0.6
cssselect2==0.7.0
Django==4.1
django-crispy-forms==1.14.0
django-filter==23.2
django-htmx==1.14.0
django-phonenumber-field==7.1.0
django-tables2==2.5.3
django-taggit==3.0.0
Faker==18.3.1
flower==2.0.0
fonttools==4.40.0
gobject==0.1.0
html5lib==1.1
humanize==4.6.0
kombu==5.3.1
numpy==1.24.3
opencv-python==4.7.0.72
phonenumberslite==8.13.11
Pillow==9.4.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tornado==6.3.2
typing_extensions==4.6.2
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
weasyprint==56.1
webencodings==0.5.1
zopfli==0.2.2
